Separate participantList reducer
