+++ src/components/ParticipantItem/ParticipantItem.js
@@ -122,8 +122,8 @@ class ParticipantItem extends Component {
 }
 
 const mapStateToProps = state => ({
-  noOfMember: state.participanti.noOfMember,
-  data: state.participanti.participants
+  noOfMember: state.participantList.noOfMember,
+  data: state.participantList.participants
 });
 
 export default connect(mapStateToProps)(ParticipantItem);
